backend/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.0
starlette==0.46.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1

# HTTP and networking
aiohttp==3.11.16
httpx==0.28.1
requests==2.32.3
websockets==14.2
httpcore==1.0.7
h11==0.14.0
h2==4.2.0

# OpenAI
openai==1.101.0
openai-agents==0.2.8
tiktoken==0.8.0


# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0


python-dotenv==1.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
tqdm==4.67.1
click==8.1.8
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock>=3.14.0
pytest-asyncio==0.21.1

# Type hints and inspection
typing_extensions==4.13.1
typing-inspection==0.4.0
annotated-types==0.7.0
types-requests==2.32.0.20250328

# Other
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
multidict==6.4.2
packaging==24.2
urllib3==2.3.0
yarl==1.19.0


# Linters and formatters
ruff>=0.12.0
black>=23.0.0

# Test optimization
pytest-xdist>=3.5.0

# Claude Code 
claude-code-sdk